requirements.txt
# Core ML/DL frameworks
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.16.0
safetensors>=0.4.0

# Training frameworks
trl>=0.7.10
peft>=0.7.0
bitsandbytes>=0.41.0

# Tokenization
sentencepiece>=0.1.99
protobuf>=3.20.0

# HuggingFace integration
huggingface_hub>=0.20.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Financial data
yfinance>=0.2.35
requests>=2.31.0

# ML utilities
scikit-learn>=1.3.0

# Configuration and environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Progress and logging
tqdm>=4.66.0
colorlog>=6.8.0

# Monitoring (optional)
wandb>=0.16.0
tensorboard>=2.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# DeepSpeed (for multi-GPU training)
deepspeed>=0.12.0

# Utilities
python-dateutil>=2.8.2
